requirements.txt
google-generativeai>=0.3.0
streamlit>=1.28.0
streamlit-option-menu>=0.3.2
plotly>=5.17.0
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
